pyproject.toml
[project]
name = "iut-market-oracle"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=42.0",
    "fastapi>=0.110",
    "matplotlib>=3.10.8",
    "numpy>=1.26",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "scipy>=1.15.3",
    "uvicorn[standard]>=0.27",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27",
    "pytest>=8.0",
    "ruff>=0.4",
    "pyright>=1.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
# sys.path manipulation must precede matplotlib.use("Agg") and all other imports
"src/report.py" = ["E402"]
# sys.path manipulation must precede all intra-project imports
"src/api/app.py" = ["E402"]

[tool.pyright]
pythonVersion = "3.10"
include = ["src", "tests"]
strict = []

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pyright>=1.1.408",
    "pytest>=9.0.3",
    "ruff>=0.15.11",
]
